refactor: Spark 프레임워크에 사용되던 클래스 제거
