+++ src/main/java/wooteco/chess/WebApplication.java
@@ -4,8 +4,8 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class SpringChessApplication {
+public class WebApplication {
 	public static void main(String[] args) {
-		SpringApplication.run(SpringChessApplication.class, args);
+		SpringApplication.run(WebApplication.class, args);
 	}
 }
