--- src/main/resources/templates/index.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <link href="/index.css" rel="stylesheet" type="text/css">
-    <title>체스</title>
-</head>
-<body>
-<div class="container">
-    <div class="room-container">
-    </div>
-    <div class="side">
-        <form action="/join" method="post">
-            <input name="room-name" placeholder="방 이름을 입력해주세요." type="text">
-            <input type="submit" value="입장">
-        </form>
-        <br/>
-        <form action="/create" method="post">
-            <input name="room-name" placeholder="방 이름을 입력해주세요." type="text">
-            <input type="submit" value="방 만들기">
-        </form>
-        <br/>
-        <br/>
-        {{#error}}
-        <label>{{.}}</label>
-        {{/error}}
-    </div>
-</div>
-<script src="/index.js"></script>
-</body>
-</html>
