--- src/test/java/wooteco/chess/ChessApplicationTests.java
@@ -1,13 +1,0 @@
-package wooteco.chess;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ChessApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
